+++ assets/galletcity_tiles.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="galletcity_tiles" tilewidth="8" tileheight="8" tilecount="168" columns="8">
+ <image source="galletcity_tiles.png" width="64" height="168"/>
+</tileset>
