+++ assets/test.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="punyworld-overworld-tileset" tilewidth="16" tileheight="16" tilecount="1755" columns="27">
+ <image source="../../../../Users/vicartas/Downloads/punyworld-overworld-tileset.png" width="432" height="1040"/>
+</tileset>
