+++ src/types/tiled.types.ts
@@ -1,35 +1,10 @@
-export type TilEdTileset = {
-    $: TilEdTileSetData,
-    image: TilEdImages[]
-}
-
-export type TilEdTileSetData = {
-    columns: string;
-    name: string;
-    tilecount: string;
-    tiledversion: string;
-    tileheight: string;
-    tilewidth: string;
-    version: string;
-}
-
-export type TilEdImages = {
-    $: TilEdImageData;
-}
-
-export type TilEdImageData = {
-    height: string;
-    source: string;
-    width: string;
-}
-
 export type TilEdMap = {
-    $: TilEdMapData;
+    $: TilEdMapMetaData;
+    tileset: TilEdTileset[];
     layer: TilEdLayer[];
-    tileset: TilEdInlineTileset[];
 }
 
-export type TilEdMapData = {
+export type TilEdMapMetaData = {
     version: string;
     tiledversion: string;
     orientation: string;
@@ -43,87 +18,49 @@ export type TilEdMapData = {
     nextlayerid: string;
 }
 
-export type TilEdLayer = {
-    $: TilEdLayerData;
-    data: TilEdLayerTiles[];
+export type TilEdTileset = {
+    $: TilEdTileSetMetaData,
+    image: TilEdImage[]
 }
 
-export type TilEdLayerData = {
+export type TilEdTileSetMetaData = {
+    firstgid: string;
+    name: string;
+    tilewidth: string;
+    tileheight: string;
+    tilecount: string;
+    columns: string;
+}
+
+export type TilEdImage = {
+    $: TilEdImageMetaData;
+}
+
+export type TilEdImageMetaData = {
+    height: string;
+    source: string;
+    width: string;
+}
+
+export type TilEdLayer = {
+    $: TilEdLayerMetaData;
+    data: TilEdLayerData[];
+}
+
+export type TilEdLayerMetaData = {
     id: string;
     name: string;
     height: string;
     width: string;
 }
 
-export type TilEdLayerTiles = {
+export type TilEdLayerData = {
     _: string;
-    $: TilEdLayerTilesData;
+    $: TilEdLayerDataMetaData;
 }
 
-export type TilEdLayerTilesData = {
+export type TilEdLayerDataMetaData = {
     encoding: TilEdLayerEncoding;
 }
 
 export type TilEdLayerEncoding = 'csv' | 'base64';
-
-export type TilEdInlineTileset = {
-    $: TilEdInlineTileSetData;
-}
-
-export type TilEdInlineTileSetData = {
-    firstgid: string;
-    source: string;
-}
-
-// <?xml version="1.0" encoding="UTF-8"?>
-// <map version="1.10" tiledversion="1.10.2" orientation="orthogonal" renderorder="right-down" width="30" height="20" tilewidth="8" tileheight="8" infinite="0" nextlayerid="3" nextobjectid="1">
-//  <tileset firstgid="1" source="galletcity_tiles.tsx"/>
-//  <layer id="1" name="Tile Layer 1" width="30" height="20">
-//   <data encoding="csv">
-// 3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,
-// 5,5,8,33,33,4,5,5,5,5,5,5,5,5,50,50,50,50,5,5,5,5,5,5,5,5,5,5,5,5,
-// 95,95,16,11,11,28,95,95,95,95,95,95,95,95,58,58,58,58,159,95,95,95,95,95,95,95,95,95,95,95,
-// 127,127,127,128,128,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,127,
-// 47,48,135,136,136,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,49,49,49,49,135,135,
-// 71,72,32,11,11,28,55,55,55,55,55,55,55,55,55,55,55,55,55,55,55,55,55,55,57,57,57,57,159,55,
-// 102,80,32,33,33,36,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,38,39,39,39,39,37,
-// 102,72,32,33,33,33,10,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,10,3,3,3,10,3,
-// 102,80,32,33,33,33,10,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,10,33,33,33,10,3,
-// 102,72,32,33,33,4,5,5,73,75,74,75,74,3,3,46,47,47,47,47,47,47,47,48,3,8,33,33,60,46,
-// 102,80,32,11,11,7,14,4,81,7,7,7,93,8,14,54,94,55,55,55,55,55,55,56,84,24,11,11,60,54,
-// 102,72,32,9,2,28,22,28,137,7,89,82,93,16,22,62,86,63,63,63,63,63,63,64,26,32,9,2,60,62,
-// 102,80,32,33,2,28,14,20,145,146,146,146,146,24,14,54,70,102,71,71,102,71,78,72,26,32,2,2,60,54,
-// 102,72,32,33,2,28,22,20,153,154,154,154,154,32,22,62,70,102,102,78,79,102,71,80,26,32,2,17,60,62,
-// 102,80,32,33,2,28,14,20,153,154,154,154,154,32,14,54,70,102,102,71,102,102,102,72,26,32,2,2,85,54,
-// 79,72,32,33,9,28,22,28,153,154,154,154,154,32,22,62,70,79,102,102,79,78,79,80,26,32,17,2,85,62,
-// 103,103,32,11,11,7,33,28,161,154,154,162,154,32,33,54,86,87,87,87,87,87,87,56,84,32,2,2,60,54,
-// 65,65,65,66,33,28,25,36,38,39,39,39,39,40,18,62,94,95,95,95,95,95,95,64,33,32,2,9,60,62,
-// 93,93,84,26,33,28,103,103,103,103,103,104,103,103,103,103,103,103,103,103,103,103,103,103,104,51,11,11,12,103,
-// 7,7,84,26,33,36,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,40,33,33,36,38
-// </data>
-//  </layer>
-//  <layer id="2" name="Tile Layer 2" width="30" height="20" visible="0">
-//   <data encoding="csv">
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,14,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-// 0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
-// </data>
-//  </layer>
-// </map>
